refactor: rename github pkg to ghclient, inline handlers, unexport mode-specific fetchers
