+++ internal/ghclient/graphql.go
@@ -1,4 +1,4 @@
-package github
+package ghclient
 
 import (
 	"bytes"
